+++ spotbugs/src/main/java/edu/umd/cs/findbugs/detect/FindPotentialSecurityCheckBasedOnUntrustedSource.java
@@ -1,13 +1,36 @@
-package edu.umd.cs.findbugs.detect;
+/*
+ * SpotBugs - Find bugs in Java programs
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ */
+
+ package edu.umd.cs.findbugs.detect;
 
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
-
+import java.util.Stack;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.apache.bcel.Const;
+import org.apache.bcel.classfile.Attribute;
+import org.apache.bcel.classfile.BootstrapMethods;
 import org.apache.bcel.classfile.Code;
+import org.apache.bcel.classfile.ConstantInvokeDynamic;
 import org.apache.bcel.classfile.JavaClass;
 import org.apache.bcel.classfile.LocalVariable;
 import org.apache.bcel.classfile.Method;
@@ -18,13 +41,18 @@ import edu.umd.cs.findbugs.BugReporter;
 import edu.umd.cs.findbugs.OpcodeStack;
 import edu.umd.cs.findbugs.SourceLineAnnotation;
 import edu.umd.cs.findbugs.ba.AnalysisContext;
+import edu.umd.cs.findbugs.ba.XClass;
 import edu.umd.cs.findbugs.ba.XField;
 import edu.umd.cs.findbugs.ba.XMethod;
 import edu.umd.cs.findbugs.ba.ch.Subtypes2;
 import edu.umd.cs.findbugs.bcel.OpcodeStackDetector;
 import edu.umd.cs.findbugs.classfile.ClassDescriptor;
+import edu.umd.cs.findbugs.util.BootstrapMethodsUtil;
 
 public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStackDetector {
+
+    private static final Matcher NESTED_CLASS_VARIABLE_NAME_MATCHER = Pattern.compile("val\\$.*").matcher("");
+
     private static class CalleeInfo {
         ClassDescriptor calledClass;
         XMethod calledMethod;
@@ -51,6 +79,30 @@ public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStac
         }
     }
 
+    private static class LambdaInfo {
+        Method lambdaMethod;
+        String[] argumentNames;
+
+        LambdaInfo(Method method, String[] argNames) {
+            lambdaMethod = method;
+            argumentNames = argNames;
+        }
+    }
+
+    private static class LambdaCallInfo {
+        JavaClass callerClass;
+        XMethod callerMethod;
+        SourceLineAnnotation srcLine;
+        String[] argumentNames;
+
+        LambdaCallInfo(JavaClass cls, XMethod met, SourceLineAnnotation line, String[] argNames) {
+            callerClass = cls;
+            callerMethod = met;
+            srcLine = line;
+            argumentNames = argNames;
+        }
+    }
+
     private static class CallPair {
         CalleeInfo outside;
         CallerInfo inside;
@@ -63,11 +115,16 @@ public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStac
 
     private Map<XMethod, Set<CalleeInfo>> nonFinalMethodsCalledOnParam = new HashMap<>();
     private Map<XMethod, Set<CallerInfo>> methodsCalledInsidePrivilegedAction = new HashMap<>();
+    private Map<OpcodeStack.Item, LambdaInfo> lambdaFunctions = new HashMap<>();
+    private Map<Method, LambdaCallInfo> lambdaCalledInDoPrivileged = new HashMap<>();
 
-    private String topParameterName = null;
+    private Stack<String> parameterNameStack = new Stack<>();
+
+    private LambdaInfo currentLambda = null;
 
     private boolean isDoPrivileged = false;
     private boolean isDoPrivilegedRun = false;
+    private boolean isLambdaCalledInDoPrivileged = false;
 
     BugAccumulator bugAccumulator;
 
@@ -77,19 +134,22 @@ public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStac
 
     @Override
     public void visit(JavaClass obj) {
+        nonFinalMethodsCalledOnParam.clear();
         isDoPrivileged = Subtypes2.instanceOf(getDottedClassName(), "java.security.PrivilegedAction")
                 || Subtypes2.instanceOf(getDottedClassName(), "java.security.PrivilegedExceptionAction");
     }
 
     @Override
     public void visit(Method obj) {
-        nonFinalMethodsCalledOnParam.clear();
-        isDoPrivilegedRun = isDoPrivileged && "run".equals(getMethodName());
+        isDoPrivilegedRun = isDoPrivileged && "run".equals(getMethodName())
+                && getMethodSig().startsWith("()");
+        isLambdaCalledInDoPrivileged = lambdaCalledInDoPrivileged.containsKey(obj);
     }
 
     @Override
     public void visit(Code obj) {
-        if (!isDoPrivilegedRun && (!getThisClass().isPublic() || !getMethod().isPublic())) {
+        if (!isDoPrivilegedRun && !isLambdaCalledInDoPrivileged &&
+                (!getThisClass().isPublic() || !getMethod().isPublic())) {
             return;
         }
         super.visit(obj);
@@ -102,7 +162,20 @@ public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStac
 
     @Override
     public void sawOpcode(int seen) {
-        if (seen == Const.INVOKEVIRTUAL && getXClassOperand() != null && getXMethodOperand() != null) {
+        currentLambda = null;
+
+        if (seen == Const.INVOKEDYNAMIC) {
+            ConstantInvokeDynamic constDyn = (ConstantInvokeDynamic) getConstantRefOperand();
+            for (Attribute attr : getThisClass().getAttributes()) {
+                if (attr instanceof BootstrapMethods) {
+                    Method method = BootstrapMethodsUtil.getMethodFromBootstrap((BootstrapMethods) attr,
+                            constDyn.getBootstrapMethodAttrIndex(), getConstantPool(), getThisClass());
+                    String[] paramNames = getParamNames();
+                    currentLambda = new LambdaInfo(method, paramNames);
+                    break;
+                }
+            }
+        } else if (seen == Const.INVOKEVIRTUAL && getXClassOperand() != null && getXMethodOperand() != null) {
             if (getXMethodOperand().getSignature().endsWith("V")) {
                 return;
             }
@@ -116,9 +189,24 @@ public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStac
                 return;
             }
 
+            if (isLambdaCalledInDoPrivileged) {
+                LambdaCallInfo lambdaCall = lambdaCalledInDoPrivileged.get(getMethod());
+                if (!getXClassOperand().isFinal() && !getXMethodOperand().isFinal() && object.isInitialParameter()
+                        && isLambdaNestingMethodLocalVariable(object, lambdaCall)) {
+                    CalleeInfo callBefore = lookForCalledOutside(lambdaCall.callerClass, lambdaCall.callerMethod,
+                            getXClassOperand(), getXMethodOperand(),
+                            lambdaCall.argumentNames[object.getRegisterNumber()]);
+                    if (callBefore != null) {
+                        reportBug(lambdaCall.callerClass, lambdaCall.callerMethod, lambdaCall.srcLine, callBefore,
+                                SourceLineAnnotation.fromVisitedInstruction(this));
+                    }
+                }
+                return;
+            }
+
             if (getXClass().isPublic() && getXMethod().isPublic() && !getXClassOperand().isFinal()
                     && !getXMethodOperand().isFinal() && object.isInitialParameter() && object.getXField() == null
-                    && topParameterName != null) {
+                    && !parameterNameStack.empty() && parameterNameStack.peek() != null) {
                 addTononFinalMethodsCalledOnParam(getClassDescriptorOperand(), getXMethodOperand(), object);
             }
         } else if (seen == Const.INVOKESTATIC && getXMethodOperand() != null
@@ -127,8 +215,26 @@ public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStac
             CallPair callPair = lookForCalledOutsideAndInside(action);
             if (callPair != null) {
                 reportBug(callPair);
+                return;
+            }
+
+            LambdaInfo lambda = lambdaFunctions.get(action);
+            if (lambda != null) {
+                lambdaCalledInDoPrivileged.put(lambda.lambdaMethod, new LambdaCallInfo(getThisClass(), getXMethod(),
+                        SourceLineAnnotation.fromVisitedInstruction(this), lambda.argumentNames));
             }
         }
+    }
+
+    private String[] getParamNames() {
+        String[] names = new String[stack.getStackDepth()];
+        for (int i = 0; i < stack.getStackDepth() && i < parameterNameStack.size(); ++i) {
+            OpcodeStack.Item param = stack.getStackItem(i);
+            if (param.isInitialParameter() && param.getXField() == null) {
+                names[i] = parameterNameStack.pop();
+            }
+        }
+        return names;
     }
 
     private boolean isNestingMethodLocalVariable(OpcodeStack.Item object) {
@@ -137,27 +243,29 @@ public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStac
             return false;
         }
 
-        return field.getName().matches("val\\$.*");
+        return NESTED_CLASS_VARIABLE_NAME_MATCHER.reset(field.getName()).matches();
+    }
+
+    private boolean isLambdaNestingMethodLocalVariable(OpcodeStack.Item object, LambdaCallInfo lambdaCall) {
+        if (object.getRegisterNumber() < 0) {
+            return false;
+        }
+
+        return object.getRegisterNumber() <= lambdaCall.argumentNames.length
+                && lambdaCall.argumentNames[object.getRegisterNumber()] != null;
     }
 
     private void addToMethodsCalledInsidePrivilegedAction(XMethod calledMethod, OpcodeStack.Item object) {
-        Set<CallerInfo> objects = methodsCalledInsidePrivilegedAction.get(calledMethod);
-        if (objects == null) {
-            objects = new HashSet<>();
-            methodsCalledInsidePrivilegedAction.put(calledMethod, objects);
-        }
+        Set<CallerInfo> objects = methodsCalledInsidePrivilegedAction.computeIfAbsent(calledMethod,
+                k -> new HashSet<>());
         objects.add(new CallerInfo(object, getThisClass(),
                 SourceLineAnnotation.fromVisitedInstruction(this)));
     }
 
     private void addTononFinalMethodsCalledOnParam(ClassDescriptor calledClass, XMethod calledMethod,
             OpcodeStack.Item object) {
-        Set<CalleeInfo> objects = nonFinalMethodsCalledOnParam.get(getXMethod());
-        if (objects == null) {
-            objects = new HashSet<>();
-            nonFinalMethodsCalledOnParam.put(getXMethod(), objects);
-        }
-        objects.add(new CalleeInfo(calledClass, calledMethod, topParameterName,
+        Set<CalleeInfo> objects = nonFinalMethodsCalledOnParam.computeIfAbsent(getXMethod(), k -> new HashSet<>());
+        objects.add(new CalleeInfo(calledClass, calledMethod, parameterNameStack.peek(),
                 SourceLineAnnotation.fromVisitedInstruction(this)));
     }
 
@@ -190,6 +298,22 @@ public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStac
         return null;
     }
 
+    private CalleeInfo lookForCalledOutside(JavaClass callerClass, XMethod callerMethod, XClass calledClass,
+            XMethod calledMethod, String argumentName) {
+        Set<CalleeInfo> callees = nonFinalMethodsCalledOnParam.get(callerMethod);
+        if (callees == null) {
+            return null;
+        }
+
+        for (CalleeInfo calleeInfo : callees) {
+            if (calleeInfo.calledMethod == calledMethod && calleeInfo.calledOnName.equals(argumentName)) {
+                return calleeInfo;
+            }
+        }
+
+        return null;
+    }
+
     private boolean isTheSame(CallerInfo inside, CalleeInfo outside, OpcodeStack.Item action) {
         XField field = inside.calledOn.getXField();
         if (field == null) {
@@ -218,9 +342,34 @@ public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStac
                         .addSourceLine(callPair.inside.srcLine), this);
     }
 
+    private void reportBug(JavaClass cls, XMethod method, SourceLineAnnotation srcLine,
+            CalleeInfo calleInfo, SourceLineAnnotation insideSrcLine) {
+        bugAccumulator.accumulateBug(new BugInstance(this, "USC_POTENTIAL_SECURITY_CHECK_BASED_ON_UNTRUSED_SOURCE",
+                NORMAL_PRIORITY)
+                        .addClass(cls)
+                        .addMethod(method)
+                        .addSourceLine(srcLine)
+                        .addClass(calleInfo.calledClass.getClassName())
+                        .addCalledMethod(calleInfo.calledClass.getClassName(),
+                                calleInfo.calledMethod.getName(), calleInfo.calledMethod.getSignature(),
+                                calleInfo.calledMethod.isStatic())
+                        .addSourceLine(calleInfo.srcLine)
+                        .addSourceLine(insideSrcLine), this);
+    }
+
     @Override
     public void afterOpcode(int seen) {
         super.afterOpcode(seen);
+
+        if (currentLambda != null && stack.getStackDepth() > 0) {
+            lambdaFunctions.put(stack.getStackItem(0), currentLambda);
+            return;
+        }
+
+        if (seen != Const.ALOAD && seen != Const.ALOAD_0 && seen != Const.ALOAD_1 && seen != Const.ALOAD_2
+                && seen != Const.ALOAD_3) {
+            parameterNameStack.clear();
+        }
 
         LocalVariable localVar = null;
         if (seen == Const.ALOAD_1 && getXMethod().getNumParams() >= 1) {
@@ -234,9 +383,10 @@ public class FindPotentialSecurityCheckBasedOnUntrustedSource extends OpcodeStac
         }
 
         if (localVar == null) {
+            parameterNameStack.push(null);
             return;
         }
 
-        topParameterName = localVar.getName();
+        parameterNameStack.push(localVar.getName());
     }
 }